오전 작업
보스 - 네트워크 오브젝트 적용 / 스크립트 수정
bomber 탄 원래대로
유니크 아이템 오브젝트 리워크

+++ Final Project/Assets/Attacker/Scripts/PlayerController.cs
@@ -318,7 +318,7 @@ public class PlayerController : NetworkBehaviour
         }
     }
 
-    public void hit(int damaged)
+    public void hit(float damaged)
     {
         hp -= damaged;
     }
